Moved ClusterType to from clusters.py to lib.py
